Add the WiX3 fragment that SOC uses and delete unused .wxi
As far as I can tell, from a GitHub search, nobody uses the .wxi files
that were here. There are copies of SOC's version, and it works, so lets
use that for everyone. This drops support for the single bitness
version, so if we ever need that it'll need to be added.